LSF support, minor bug/doc fixes.

+++ kmer/ESTmapper/util/2-search.pl
@@ -1,48 +1,58 @@
-sub countTheCountFile {
-    my ($path, $s) = @_;
-    my $c = 0;
+use strict;
 
-    #  If the hits file is NOT found, remove the count file.
-    #
-    if (! -e "$path/1-search/$s.hits") {
-        unlink "$path/1-search/$s.count";
+sub checkFinished {
+    my ($path, $cdnaInInput, @scafList) = @_;
+    my @searchesToRun;
+
+    open(S, "> $path/1-search/run.sh");
+    foreach my $s (@scafList) {
+        my $c = 0;
+
+        #  If the hits file is NOT found, remove the count file.
+        #
+        unlink "$path/1-search/$s.count" if (! -e "$path/1-search/$s.hits");
+
+        #  If there is a count file, count the number of lines in it.
+        #
+        $c = int(`wc -l < $path/1-search/$s.count`) if (-e "$path/1-search/$s.count");
+
+        #  There should be one for each cDNA.
+        #
+        if ($c != $cdnaInInput) {
+            print STDERR "ESTmapper/search-- Found signal for $c cDNA out of $cdnaInInput for segment $s.\n";
+
+            print S "$path/1-search/$s.cmd";
+            push @searchesToRun, "$path/1-search/$s.cmd";
+        }
     }
+    close(S);
 
-    #  If there is a count file, count the number of lines in it.
-    #  There should be one for each cDNA.
-    #
-    if (-e "$path/1-search/$s.count") {
-        $c = int(`wc -l < $path/1-search/$s.count`);
-    }
-
-    print STDERR "ESTmapper/search-- found $c cDNA for search $s\n";
-
-    return($c);
+    return(@searchesToRun);
 }
-
 
 
 sub search {
     my $startTime = time();
-    my $errHdr       = "ERROR: ESTmapper/search--";
     my @ARGS         = @_;
     my $path         = "";
     my $opts         = "";
-    my $cdna         = "";
     my $mersize      = 20;
     my $maskFile     = "";
     my $builddir     = undef;
     my $buildprefix  = undef;
     my $verbose      = "";
     my $stats        = 1;
-    my $farm         = 0;
-    my $farmqueue    = "";
-    my $farmcode     = "";
-    my $local        = 1;
+
+    my $farmname;
+    my $farmcode;
+    my $farmqueue;
+    my $filtqueue;
+
+    my $hitMemory    = 600;    #  Don't change the value without 3-filter
+
     my $numthread    = 2;
     my $numproc      = 4;
     my $runnow       = 1;
-    my $maxintron    = "-maxintron 2000000";
     my $species      = "";
 
     print STDERR "ESTmapper: Performing a search.\n";
@@ -50,67 +60,47 @@ sub search {
     while (scalar @ARGS > 0) {
         my $arg = shift @ARGS;
 
+        $verbose = "-verbose" if ($arg eq "-verbose");
+
         if ($arg eq "-searchest") {
             $path = shift @ARGS;
-            $opts = "-singlelength 20 -multiplelength 30 -smallsequence 100";
+            $opts = "-maxintron 2000000 -singlelength 20 -multiplelength 30 -smallsequence 100";
         }
         if ($arg eq "-searchmrna") {
             $path = shift @ARGS;
-            $opts = "-singlelength 30 -multiplelength 50 -smallsequence 0";
+            $opts = "-maxintron 2000000 -singlelength 30 -multiplelength 50 -smallsequence 0";
         }
         if ($arg eq "-searchsnp") {
             $path = shift @ARGS;
-            $opts = "-singlecoverage 0.3 -multiplecoverage 0.3 -smallsequence 10000000 -extendminimum 100 -extendweight 2";
+            $opts = "-maxintron 2000000 -singlecoverage 0.3 -multiplecoverage 0.3 -smallsequence 10000000 -extendminimum 100 -extendweight 2";
         }
         if ($arg eq "-searchopts") {
             $opts .= " " . shift @ARGS;
         }
-        if ($arg eq "-cdna") {
-            $cdna = shift @ARGS;
-        }
         if ($arg eq "-mersize") {
             $mersize = shift @ARGS;
         }
-        if ($arg eq "-maskmers") {
-            $species  = "";
-            $maskFile = shift @ARGS;
-        }
+
         if ($arg eq "-prebuild") {
             $builddir    = shift @ARGS;
             $buildprefix = shift @ARGS;
         }
-        if ($arg eq "-nomaskmers") {
-            $maskFile = "";
-        }
-        if ($arg eq "-verbose") {
-            $verbose = "-verbose";
-        }
-        if ($arg eq "-stats") {
-            $stats = 1;
-        }
-        if ($arg eq "-nostats") {
-            $stats = 0;
-        }
-        if ($arg eq "-farmsearches") {
-            $farm     = 1;
-            $local    = 0;
-            $farmqueue = shift @ARGS;
-            $farmcode  = shift @ARGS
-        }
-        if ($arg eq "-localsearches") {
-            $farm  = 0;
-            $local = 1;
-            $numproc = shift @ARGS;
-        }
-        if ($arg eq "-searchthreads") {
-            $numthread = shift @ARGS;
-        }
-        if ($arg eq "-runlater") {
-            $runnow = 0;
-        }
-        if ($arg eq "-maxintron") {
-            $maxintron = "-maxintron " . shift @ARGS;
-        }
+
+        $stats = 1 if ($arg eq "-stats");
+        $stats = 0 if ($arg eq "-nostats");
+
+        $farmname  = shift @ARGS if ($arg eq "-lsfjobname");
+        $farmcode  = shift @ARGS if ($arg eq "-lsfproject");
+        $farmqueue = shift @ARGS if ($arg eq "-lsfsearchqueue");
+        $filtqueue = shift @ARGS if ($arg eq "-lsffilterqueue");
+
+        $numproc   = shift @ARGS if ($arg eq "-localsearches");
+        $numthread = shift @ARGS if ($arg eq "-searchthreads");
+
+        $hitMemory = shift @ARGS if ($arg eq "-hitsortmemory");
+
+        $runnow    = 0           if ($arg eq "-runlater");
+
         if ($arg eq "-species") {
             $species  = shift @ARGS;
             $maskFile = "";
@@ -130,43 +120,26 @@ sub search {
                 $maskFile = "";
             }
         }
+
+        #  Use carefully or not at all.
+        $maskFile = shift @ARGS if ($arg eq "-maskmers");
     }
 
-    ($path eq "") and die "$errHdr no directory given.\n";
-    (! -d "$path") and die "$errHdr no directory '$path' found!\n";
-    (! -f "$path/0-input/scaffolds-list") and die "$errHdr no scaffolds-list?\n";
-    (($maskFile ne "") && (! -e "$maskFile")) and die "$errHdr Can't find -maskmers '$maskFile'!\n";
-    ($cdna eq "") and die "$errHdr Please supply some cDNA to search for!\n";
-    (! -f $cdna) and die "$errHdr cDNA sequences '$cdna' not found!\n";
-
-
-    if ($maskFile eq "") {
-        print STDERR "ESTmapper/search-- No file of maskmers - mer masking disabled.\n";
-    } else {
-        print STDERR "ESTmapper/search-- Using '$maskFile' for mer masking.\n";
-    }
-
-    system("mkdir $path/1-search") if (! -d "$path/1-search");
-
+    #  If we're all done, skip the searches.
+    #
     if (-e "$path/1-search/allDone") {
         print STDERR "ESTmapper/search-- this stage all done, and results checked previously.\n";
         return;
     }
 
-    #  Save a pointer to the cDNA used
-    #
-    if (! -e "$path/0-input/cDNA.fasta") {
-        system("ln -s $cdna      $path/0-input/cDNA.fasta");
-    }
+    ($path eq "")                              and die "ERROR: ESTmapper/search-- no directory given.\n";
+    (! -d "$path")                             and die "ERROR: ESTmapper/search-- no directory '$path' found!\n";
+    (! -f "$path/0-input/scaffolds-list")      and die "ERROR: ESTmapper/search-- no scaffolds-list?\n";
+    (($maskFile ne "") && (! -e "$maskFile"))  and die "ERROR: ESTmapper/search-- Can't find mer mask file '$maskFile'!\n";
 
-    if (! -e "$path/0-input/cDNA.fastaidx") {
-        if (-f "${cdna}idx") {
-            system("ln -s ${cdna}idx $path/0-input/cDNA.fastaidx");
-        } else {
-            print STDERR "Building index for cDNA sequences.\n";
-            system("$leaff -F $path/0-input/cDNA.fasta");
-        }
-    }
+    system("mkdir $path/1-search")                  if (! -d "$path/1-search");
+
+    my $cdnaInInput  = int(`$leaff -F $path/0-input/cDNA.fasta -d`);
 
     #  Read the list of segments the searches used
     #
@@ -182,7 +155,7 @@ sub search {
     #
     foreach my $s (@scafList) {
         open(F, "> $path/1-search/$s.cmd");
-        print F "$searchGENOME $verbose -binary -mersize $mersize $maxintron $opts -numthreads $numthread";
+        print F "$searchGENOME $verbose -binary -mersize $mersize $opts -numthreads $numthread";
         print F " -buildonly $builddir/$buildprefix.$mersize.$s" if (defined($buildprefix));
         print F " -cdna $path/0-input/cDNA.fasta";
         print F " -genomic $path/0-input/genomic.fasta";
@@ -195,27 +168,20 @@ sub search {
         system("chmod 755 $path/1-search/$s.cmd");
     }
 
-    my $cdnaInInput  = int(`$leaff -F $path/0-input/cDNA.fasta -d`);
+    #  We build a list of jobs to run, and put it into
+    #  $path/1-search/run.sh.
+    #
+    my @searchesToRun = checkFinished($path, $cdnaInInput, @scafList);
 
     #  Run searches.  If the search terminated properly, the
     #  hit-counts file should exist.  Run (maybe re-run) the search if
     #  it isn't there.
     #
     if ($runnow) {
-        if ($local) {
+        if (!defined($farmqueue)) {
             print STDERR "ESTmapper/search-- Local mode requested; ", scalar @scafList, " processes to compute,\n";
             print STDERR "ESTmapper/search-- Local mode requested; $numproc concurrent processes,\n";
             print STDERR "ESTmapper/search-- Local mode requested; each with $numthread threads.\n";
-
-            #  Build a list of things to run.
-            #
-            my @searchesToRun;
-
-            foreach my $s (@scafList) {
-                if (countTheCountFile($path, $s) != $cdnaInInput) {
-                    push @searchesToRun, $s;
-                }
-            }
 
             my $numTries = 0;
           again:
@@ -230,19 +196,14 @@ sub search {
                 &libBri::schedulerSetShowCommands(0);
                 &libBri::schedulerSetShowStatus(1);
                 foreach my $s (@searchesToRun) {
-                    &libBri::schedulerSubmit("sh $path/1-search/$s.cmd");
+                    &libBri::schedulerSubmit("sh $s");
                 }
                 &libBri::schedulerFinish();
 
                 #  See if anything failed.
                 #
-                undef @searchesToRun;
                 print STDERR "ESTmapper/search-- checking search output.  All should have $cdnaInInput cDNA.\n";
-                foreach my $s (@scafList) {
-                    if (countTheCountFile($path, $s) != $cdnaInInput) {
-                        push @searchesToRun, $s;
-                    }
-                }
+                @searchesToRun = checkFinished($path, $cdnaInInput, @scafList);
             }
 
             if (($numTries < 3) && (scalar(@searchesToRun) > 0)) {
@@ -253,50 +214,53 @@ sub search {
             if (scalar(@searchesToRun) > 0) {
                 print STDERR "ESTmapper/search-- Searches failed.\n";
                 foreach my $s (@searchesToRun) {
-                    print STDERR "ESTmapper/search-- Search $s failed.  Output saved as *.CRASH\n";
-                    rename "$path/1-search/$s.count", "$path/1-search/$s.count.CRASH";
-                    rename "$path/1-search/$s.hits", "$path/1-search/$s.hits.CRASH";
+                    if ($s =~ m/(\d+).cmd/) {
+                        print STDERR "ESTmapper/search-- Search $1 failed.  Output saved as *.CRASH\n";
+                        rename "$path/1-search/$1.count", "$path/1-search/$1.count.CRASH";
+                        rename "$path/1-search/$1.hits", "$path/1-search/$1.hits.CRASH";
+                    } else {
+                        print STDERR "ESTmapper/search-- Search $s failed.\n";
+                    }
                 }
             }
-        }
-
-        if ($farm) {
+        } else {
             open(F, "< $path/0-input/memoryLimit");
             my $farmMemory = <F>;
             close(F);
             chomp $farmMemory;
 
-            print STDERR "ESTmapper/search-- Farm mode requested; ", scalar @scafList, " processes to compute,\n";
-            print STDERR "ESTmapper/search-- Farm mode requested; each with $numthread threads,\n";
-            print STDERR "ESTmapper/search-- Farm mode requested; $farmMemory MB per process.\n";
+            my $cmd;
+            my $jobid;
 
-            for (my $x=10; $x>0; $x--) {
-                print STDERR "ESTmapper/search-- Abort in $x seconds to change. \r";
-                sleep(1);
-            }
-            print STDERR "ESTmapper/search-- STARTING FARM RUN!           \n";
+            if (scalar(@searchesToRun) > 0) {
+                print STDERR "ESTmapper/search-- LSF mode requested; ", scalar @searchesToRun, " processes to compute,\n";
+                print STDERR "ESTmapper/search-- LSF mode requested; each with $numthread threads,\n";
+                print STDERR "ESTmapper/search-- LSF mode requested; $farmMemory MB per process.\n";
 
-            my $jobsToRun = 0;
+                $cmd  = "bsub -q $farmqueue -P $farmcode -n $numthread -R \"span[hosts=1]select[mem>$farmMemory]rusage[mem=$farmMemory]\" -o $path/1-search/%J-%I.lsfout ";
+                $cmd .= " -J \"s$farmname\[1-" . scalar(@searchesToRun) . "\]\" ";
+                $cmd .= "$exechome/util/jobarray.pl $path search";
 
-            foreach my $s (@scafList) {
-                if (countTheCountFile($path, $s) != $cdnaInInput) {
-                    print STDERR "ESTmapper/search-- submitting search $s\n";
-                    if (`uname` =~ m/aix/i) {
-                        system("bsub -q $farmqueue -o $path/1-search/$s.stdout -R \"select[mem>$farmMemory]rusage[mem=$farmMemory]\" -P $farmcode $path/1-search/$s.cmd");
-                    } else {
-                        system("bsub -q $farmqueue -o $path/1-search/$s.stdout -R \"select[physmem>$farmMemory]rusage[physmem=$farmMemory]\" -P $farmcode $path/1-search/$s.cmd");
-                    }
-                    $jobsToRun++;
-                } else {
-                    print STDERR "ESTmapper/search-- search $s finished successfully!\n";
-                }
-            }
+                $jobid = runLSF($cmd);
 
-            #  Hang around waiting for them to finish.
-            if ($jobsToRun > 0) {
-                die "ESTmapper/search-- I don't know how to monitor LSF jobs!\nESTmapper/search-- Please restart when they're all done.\n";
+                #  Submit the filter, and make it wait for the searches, if they were submitted.
+                #
+                $cmd  = "bsub -q $filtqueue -P $farmcode -R \"select[mem>$hitMemory]rusage[mem=$hitMemory]\" -o $path/stage2.lsfout ";
+                $cmd .= " -w \"ended($jobid)\"" if (defined($jobid));
+                $cmd .= " -J f$farmname ";
+                $cmd .= " $ESTmapper -restart $path";
+
+                system($cmd);
+
+                print STDERR "ESTmapper/search-- Searches submitted.   Rest of run is on the farm.\n";
+
+                exit(0);
             }
         }
+    } else {
+        print STDERR "ESTmapper/search-- Please run the jobs in\n";
+        print STDERR "ESTmapper/search--   $path/1-search/run.sh\n";
+        exit(0);
     }
 
 
@@ -308,7 +272,7 @@ sub search {
     my $usrTime = 0;
     foreach my $s (@scafList) {
         if (! -e "$path/1-search/$s.count") {
-            print STDERR "$errHdr SEARCH $s DIED!\n";
+            print STDERR "ERROR: ESTmapper/search-- SEARCH $s DIED!\n";
             $died++;
         }
         if (-e "$path/1-search/$s.stats") {
@@ -324,8 +288,7 @@ sub search {
             close(F);
         }
     }
-    die "Searches have a problem ($died).\n" if ($died > 0);
-
+    die "ERROR: ESTmapper/search-- Searches have a problem ($died).\n" if ($died > 0);
 
     #  Rather lazy way to inform the next step (and future calls to this step) that we're all done.
     #
